more updates

+++ database/migrations/2023_03_11_234323_create_discs_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->decimal('glide')->nullable();
             $table->decimal('fade')->nullable();
             $table->decimal('turn')->nullable();
-            $table->decimal('weight')->nullable();
+            $table->integer('weight')->nullable();
             $table->timestamps();
         });
     }
